refector: AI 관련 로직 분리 및 MVC 구조분해
gitLab 서비스/컨트롤러 분리 진행
SVN에서 코드리뷰 로직 재사용을 위한 분리 작업 진행
